changes in store (fixed catalogue filters behavior when navigating with back button)
