+++ src/store/index.ts
@@ -5,6 +5,8 @@ import createPersistedState from 'vuex-persistedstate';
 import User from './modules/user.module';
 import Config from './modules/config.module';
 import Shop from './modules/shop.module';
+import Router from './modules/router.module';
+import CatalogueFilters from './modules/catalogue-filters.module';
 
 Vue.use(Vuex);
 
@@ -15,6 +17,8 @@ const store = new Vuex.Store({
     user: User,
     config: Config,
     shop: Shop,
+    router: Router,
+    catalogueFilters: CatalogueFilters,
   },
   plugins: [createPersistedState({
     // paths: ['user'],
